Create docker-compose.yml
This file holds all the download components of the media server. All of the services use container "vpn" as their network. This means that this network will only have one IP address accessible that will host all of the ports of the services. If Swag is located on a separate subnet, you will need to manually replace the container name in the proxy-conf file with the ip address of Swag's default gateway.